backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
Pillow>=10.2.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
aiofiles>=23.2.0
deepface
tf-keras
ollama
asyncpg